+++ lib/features/onboarding/presentation/widgets/buttons/onboarding_button.dart
@@ -1,0 +1,78 @@
+import 'package:flutter/material.dart';
+
+import '../../widgets/utils/vibration_util.dart';
+
+/// オンボーディング用の標準ボタン
+class OnboardingButton extends StatelessWidget {
+  final String text;
+  final VoidCallback onPressed;
+  final bool isPrimary;
+
+  const OnboardingButton({
+    super.key,
+    required this.text,
+    required this.onPressed,
+    this.isPrimary = true,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: double.infinity,
+      child: FilledButton(
+        onPressed: () {
+          // ボタン押下時にバイブレーション
+          VibrationUtil.buttonClick();
+          onPressed();
+        },
+        style: FilledButton.styleFrom(
+          backgroundColor: isPrimary
+              ? Theme.of(context).colorScheme.primary
+              : Colors.grey.shade800,
+          foregroundColor: Colors.white,
+          padding: const EdgeInsets.symmetric(vertical: 16),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(24),
+          ),
+        ),
+        child: Text(
+          text,
+          style: const TextStyle(
+            fontSize: 16,
+            fontWeight: FontWeight.bold,
+            color: Colors.white,
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+/// 戻るボタンウィジェット
+class OnboardingBackButton extends StatelessWidget {
+  final VoidCallback onPressed;
+
+  const OnboardingBackButton({
+    super.key,
+    required this.onPressed,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: IconButton(
+        icon: const Icon(Icons.arrow_back_ios_new),
+        onPressed: onPressed,
+        color: Colors.white,
+        style: IconButton.styleFrom(
+          backgroundColor: Colors.black26,
+          padding: const EdgeInsets.all(12),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(24),
+          ),
+        ),
+      ),
+    );
+  }
+}
