+++ lib/features/onboarding/presentation/widgets/cards/onboarding_cards.dart
@@ -1,0 +1,157 @@
+import 'package:flutter/material.dart';
+
+import '../utils/vibration_util.dart';
+
+/// オンボーディングのオプションカードウィジェット
+class OnboardingOptionCard extends StatelessWidget {
+  final String title;
+  final String? subtitle;
+  final bool isSelected;
+  final VoidCallback onTap;
+  final IconData? icon;
+  final bool autoNavigate;
+  final VoidCallback? onNext;
+
+  const OnboardingOptionCard({
+    super.key,
+    required this.title,
+    this.subtitle,
+    required this.isSelected,
+    required this.onTap,
+    this.icon,
+    this.autoNavigate = false,
+    this.onNext,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      elevation: 0,
+      margin: const EdgeInsets.only(bottom: 12),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(24),
+        side: BorderSide(
+          color: isSelected
+              ? Theme.of(context).colorScheme.primary
+              : Theme.of(context).dividerColor,
+          width: isSelected ? 2 : 1,
+        ),
+      ),
+      color: isSelected
+          ? Theme.of(context).colorScheme.primary.withOpacity(0.1)
+          : Theme.of(context).colorScheme.surface,
+      child: InkWell(
+        onTap: () {
+          // タップ時にバイブレーション
+          VibrationUtil.optionTap();
+
+          onTap();
+          // Auto-navigate to next step after selection if enabled
+          if (autoNavigate && onNext != null) {
+            Future.delayed(const Duration(milliseconds: 300), onNext);
+          }
+        },
+        borderRadius: BorderRadius.circular(24),
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
+          child: Row(
+            children: [
+              if (icon != null)
+                Icon(
+                  icon,
+                  color: isSelected
+                      ? Theme.of(context).colorScheme.primary
+                      : Colors.white.withOpacity(0.7),
+                  size: 24,
+                ),
+              if (icon != null) const SizedBox(width: 16),
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      title,
+                      style: TextStyle(
+                        fontWeight: FontWeight.w600,
+                        fontSize: 16,
+                        color: isSelected
+                            ? Theme.of(context).colorScheme.primary
+                            : Colors.white,
+                      ),
+                    ),
+                    if (subtitle != null)
+                      Padding(
+                        padding: const EdgeInsets.only(top: 4),
+                        child: Text(
+                          subtitle!,
+                          style: TextStyle(
+                            fontSize: 13,
+                            color: Colors.white.withOpacity(0.6),
+                          ),
+                        ),
+                      ),
+                  ],
+                ),
+              ),
+              if (isSelected)
+                Icon(
+                  Icons.check_circle,
+                  color: Theme.of(context).colorScheme.primary,
+                  size: 24,
+                ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+/// 利用者の評価を表示するウィジェット
+class OnboardingTestimonial extends StatelessWidget {
+  final String quote;
+  final String author;
+
+  const OnboardingTestimonial({
+    super.key,
+    required this.quote,
+    required this.author,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        color: Theme.of(context).colorScheme.surface,
+        borderRadius: BorderRadius.circular(24),
+        border: Border.all(
+          color: Theme.of(context).dividerColor,
+        ),
+      ),
+      child: Column(
+        children: [
+          Text(
+            quote,
+            style: TextStyle(
+              fontStyle: FontStyle.italic,
+              fontSize: 14,
+              color: Colors.white.withOpacity(0.8),
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 12),
+          Text(
+            '- $author',
+            style: TextStyle(
+              fontWeight: FontWeight.bold,
+              fontSize: 14,
+              color: Theme.of(context).colorScheme.primary,
+            ),
+            textAlign: TextAlign.right,
+          ),
+        ],
+      ),
+    );
+  }
+}
